Rename submodules under `cupyx.linalg` package
