+++ cupyx/linalg/__init__.py
@@ -1,3 +1,3 @@
 # "NOQA" to suppress flake8 warning
 from cupyx.linalg import sparse  # NOQA
-from cupyx.linalg.solve import invh  # NOQA
+from cupyx.linalg._solve import invh  # NOQA
