+++ cupyx/linalg/sparse/__init__.py
@@ -1,3 +1,3 @@
 
 # "NOQA" to suppress flake8 warning
-from cupyx.linalg.sparse.solve import lschol  # NOQA
+from cupyx.linalg.sparse._solve import lschol  # NOQA
