+++ tests/cupyx_tests/linalg_tests/test_solve.py
@@ -72,7 +72,7 @@ class TestXFailBatchedInvh(unittest.TestCase):
         a = self._create_symmetric_matrix(self.shape, self.dtype)
         with cupyx.errstate(linalg='ignore'):
             with self.assertRaises(cupy.cuda.cusolver.CUSOLVERError):
-                cupyx.linalg.solve._batched_invh(a)
+                cupyx.linalg._solve._batched_invh(a)
 
     def _create_symmetric_matrix(self, shape, dtype):
         a = testing.shaped_random(shape, cupy, dtype, scale=1)
